requirements_new.txt
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
appnope==0.1.4
asttokens==3.0.0
attributee==0.1.9
backcall==0.2.0
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1695989934239/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1723018415758/work
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.1.1
crcmod==1.7
cryptography==45.0.3
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dotty-dict==1.3.1
executing==2.2.0
filelock==3.14.0
fonttools==4.57.0
gmpy2 @ file:///Users/runner/miniforge3/conda-bld/gmpy2_1715527349377/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1634280454336/work
hpack==4.0.0
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1619110129307/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1726459485162/work
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1715127149914/work
jmespath==0.10.0
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.7
lap==0.5.12
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1706899993502/work
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mmcls==0.25.0
mmcv-full==1.7.2
mmdet==2.28.2
mmengine==0.10.7
mmtrack==0.14.0
model-index==0.1.11
motmetrics==1.4.0
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1678228039184/work
nest-asyncio==1.6.0
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1680692919326/work
numpy==1.22.4
opencv-python==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==1.3.5
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1719903577696/work
platformdirs==4.3.6
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycocotools==2.0.7
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1711811537435/work
pycryptodome==3.23.0
Pygments==2.19.1
pyparsing==3.1.4
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.9.0.post0
pytz==2023.4
PyYAML @ file:///Users/runner/miniforge3/conda-bld/pyyaml_1723018283744/work
pyzmq==26.4.0
requests==2.28.2
rich==13.4.2
scikit-learn==1.3.2
scipy==1.7.3
seaborn==0.13.2
shapely==2.0.7
six==1.17.0
stack-data==0.6.3
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1728484478345/work
tabulate==0.9.0
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tomli==2.2.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.4.2
tqdm==4.65.2
traitlets==5.14.3
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.13
xmltodict==0.14.2
yapf==0.43.0
zipp==3.20.2
zstandard @ file:///Users/runner/miniforge3/conda-bld/zstandard_1667296220480/work
